+++ gulp/config.js
@@ -32,7 +32,7 @@ module.exports = {
   },
   markup: {
     src: [
-      demoSrc + "/*/*"
+      demoSrc + "/**/*"
     ],
     base: './demos',
     dest: dest
@@ -44,9 +44,9 @@ module.exports = {
     outputName: 'moondash.css',
     dest: dest
   },
-  partials: {
+  templates: {
     src: [
-      src + '/**/*.partial.html'
+      src + '/**/templates/*.html'
     ],
     vendors: {
       src: [