+++ gulp/tasks/build.js
@@ -5,7 +5,7 @@ gulp.task('build', [
     'markup',
     'sass',
     'icons',
-    'partials',
+    'templates',
     'vendors',
     'test'
 ]);